+++ MelonLoader/Lemons/Cryptography/LemonSHA256.cs
@@ -1,0 +1,28 @@
+﻿using System.Security.Cryptography;
+
+namespace MelonLoader.Lemons.Cryptography
+{
+	// Modified Version of System.Security.Cryptography.SHA256 from .NET Framework's mscorlib.dll
+	public abstract class LemonSHA256 : HashAlgorithm
+	{
+		/// <summary>Initializes a new instance of <see cref="T:System.Security.Cryptography.SHA256" />.</summary>
+		protected LemonSHA256()
+			=> HashSizeValue = 256;
+
+		/// <summary>Creates an instance of the default implementation of <see cref="T:System.Security.Cryptography.SHA256" />.</summary>
+		/// <returns>A new instance of <see cref="T:System.Security.Cryptography.SHA256" />.</returns>
+		/// <exception cref="T:System.Reflection.TargetInvocationException">The algorithm was used with Federal Information Processing Standards (FIPS) mode enabled, but is not FIPS compatible.</exception>
+		/// <PermissionSet>
+		///   <IPermission class="System.Security.Permissions.SecurityPermission, mscorlib, Version=2.0.3600.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" version="1" Flags="UnmanagedCode" />
+		/// </PermissionSet>
+		public new static LemonSHA256 Create()
+			=> Create("System.Security.Cryptography.SHA256");
+
+		/// <summary>Creates an instance of a specified implementation of <see cref="T:System.Security.Cryptography.SHA256" />.</summary>
+		/// <returns>A new instance of <see cref="T:System.Security.Cryptography.SHA256" /> using the specified implementation.</returns>
+		/// <param name="hashName">The name of the specific implementation of <see cref="T:System.Security.Cryptography.SHA256" /> to be used. </param>
+		/// <exception cref="T:System.Reflection.TargetInvocationException">The algorithm described by the <paramref name="hashName" /> parameter was used with Federal Information Processing Standards (FIPS) mode enabled, but is not FIPS compatible.</exception>
+		public new static LemonSHA256 Create(string hashName)
+			=> CryptoConfig.CreateFromName(hashName) as LemonSHA256;
+	}
+}