+++ CHANGELOG.md
@@ -88,6 +88,8 @@
 56. Fixed MissingMethodException Issue with HarmonyLib.AccessTools.
 57. Fixed MissingMethodException Issue with MelonUtils.SetCurrentDomainBaseDirectory.
 58. Added --melonloader.agfregex Launch Option to Force Regex used for Assembly Generation.
+59. Added LemonSHA256 because System.Security.Cryptography.SHA256 can have Missing Methods.
+60. Fixed MissingMethodException Issue with MelonHandler.sha256.
 
 ---